Controller: Add -s option to snoop by channel name or unique ID.
Adds a new command line option boolean '-s' to change the meaning of
<dialstring> from originating a call to a new destination, to instead
snooping on an existing channel, either by channel name or unique ID.

For example...

$ bin/ari-transcriber -s 'PJSIP/7001-00000001'

Big thanks to Voisonics for sponsoring this patch.

Resolves: #21

UserNote: The ari-transcriber now takes an additional argument '-s' on
the command line to allow snooping on any Asterisk channel.